+++ apps/frontend/app/(website)/studio/main/PaneLayout/index.ts
@@ -1,0 +1,3 @@
+export * from "./tabsData";
+export * from "./AssistantTabs";
+export * from "./SignInRequired";
